Decoupling MonthByWeekFragment from its loader.
The idea is to be able to pass a custom loader to work with different sets of events. This way, we may not even need to access the calendar ContentProvider and create Event instances from other sources using a custom EventLoader.

Current implementation hard-codes the DefaultEventLoader, forcing the original behavior to work as before.